ARM: shmobile: check if cpu is allowed on power on
To decide if we can boot cpus that are not part of the
boot cluster, we will check if cci and mcpm were correcly enabled
and probed. But mcpm_init is called with an early_init call wich is
later than apmu_parse_cfg.

So, check if a cpu should be allowed to boot direcly from the
the power_on function as this will allow us to check if mcpm
and cci are enabled.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>