PM / Domains: introduce pstates
this allows to attach a series of pstates to a particular
power domain.

we can declare the number of pstates of a power domain
using devicetree pstate = <n> where n is the number of
pstastes for this domain.

if the parent domain pstates dont map "one-to-one" to
the domain pstates we can use pstate-map = < a b c ..>;
where "a b c .." is the corresponding required
state for the parent. ex.
	pstate 0 sets the parent to pstate a,
	pstste 1 sets the parent to pstate b,
	[...]

example power domain declaration:

pm-domains {
	compatible = "fake,fake-pmdomain";
	pd_a1: a1 {
		#power-domain-cells = <0>;
		pstates = <3>;

		pd_a2: a2@0 {
			#power-domain-cells = <0>;
			pstates = <3>;
			pstate-map = <0 1 1>;
			pd_a4: a4@0 {
				#power-domain-cells = <0>;
				pstates = <3>;
			};

		};

		pd_a3: a3@1 {
			#power-domain-cells = <0>;
			pstates = <3>;
		};
	};

	pd_b1: b1 {
		#power-domain-cells = <0>;
		pd_b2: b2@0 {
			#power-domain-cells = <0>;
		};

		pd_b3: b3@1 {
			#power-domain-cells = <0>;
		};
	};

};

to add pstates to a power domain, one should call
pm_genpd_add_pstates(genpd, parent, np);

when a device is attached to a power domain,
the power domains of parents will be notified of
a pstate change request, and the pstate callback
will be called for all power domains and devices
that share the same domain from "top to bottom".

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>