arm64: juno: add fake driver to cluster power domain
this shall help investigate how to manage device constraints in
power domains mixing CPU and non-CPU devices. The goal is to
allow the genpd governor to optimize the suspend state of the
cluster based on c-states chosen by menu governor and device
constraints. Based on fake driver from Axel Haslam.

Signed-off-by: Marc Titinger <mtitinger+renesas@baylibre.com>